// src/types/sanity.d.ts
export type SanitySlug = {
  _type: "slug"
  _id: string
  _key: string
  current: string
}

//                ___   __
//               /   | / /_____  ____ ___  _____
//              / /| |/ __/ __ \/ __ `__ \/ ___/
//             / ___ / /_/ /_/ / / / / / (__  )
//            /_/  |_\__/\____/_/ /_/ /_/____/

export type SanityLocale = SupportedLanguage["id"]

export type SanityLocaleString = {
  [x in SanityLocale]?: string
} & {
  // _type: 'localeString'
  // _id: string
}

export type SanityLocaleText = {
  _type: "locale_text"
  _id: string
} & { [x in SanityLocale]?: string }

export type SanityLocaleNumber = {
  _type: "locale_number"
  _id: string
} & { [x in SanityLocale]?: number }

export type SanityImage = {
  _type: "image"
  _id: string
  _key: string
  asset: {
    _type: "reference"
    _id: string
    _ref: string
  }
}

export type SanityPhoto = {
  asset: {
    _type: "reference"
    _id: string
    _ref: string
  }
  _key: string
  _type: "photo"
  _id: string
  alt?: SanityLocaleString
}

export type SanityPrice = {
  _type: "price"
  _id: string
  _key: string
  currency_symbol?: SanityLocaleString
  initial_price?: SanityLocaleNumber
  discounted_price?: SanityLocaleNumber
}
// ...
export type SanityLinkButton = {
  _type: "link_button"
  _id: string
  _key: string
  type?: "primary" | "secondary"
  label?: SanityLocaleString
  url?: string
  icon?: SanityIcon
}

export type SanityMetaData = {
  _type: "meta_data"
  _id: string
  _key: string
  meta_title?: SanityLocaleString
  meta_description?: SanityLocaleString
  meta_image?: SanityImage
}

export type SanityPromoBanner = {
  _type: "promo_banner"
  _id: string
  _key: string
  text?: SanityLocaleString
  link?: SanityLink
}

export type SanityTimeline = {
  _type: "timeline"
  _id: string
  _key: string
  start_date?: string
  end_date?: string
}

export type SanityTourTimeline = {
  _type: "tour_timeline"
  _id: string
  _key: string
  start_day?: "sun" | "mon" | "tue" | "wed" | "thu" | "fri" | "sat"
  duration?: number
}
// ...
export type SanityFeaturedToursSection = {
  _type: "featured_tours_section"
  _id: string
  _key: string
  tagline?: SanityLocaleString
  title?: SanityLocaleString
  tour_cards?: {
    badge_content?: SanityLocaleString
    content?: {
      _type: "reference"
      _id: string
      _ref: string
    } & SanityTourPage
  }[]
}
// ...
export type SanityDestinationPage = {
  _type: "destination_page"
  _id: string
  name?: SanityLocaleString
  slug?: SanitySlug
  meta_data?: SanityMetaData
  promo_banner?: SanityPromoBanner
  // hero_section?: {
  //   title?: SanityLocaleString
  //   image?: SanityImage
  // }
  discounts_section?: {
    header_section: SanityImageHeaderSection
  }
  sections?: SanityDestinationSection[]
}
// ...
export type SanityTourPageSection =
  | SanityContentSection
  | SanityFeatureSection
  | SanityReviewsSection
  | SanityPricingSection
  | SanityFAQSection
  | SanityAccommodationSection
  | SanityGallerySection
  | SanityItinerarySection
  | SanityMemorableExperiencesSection
  | SanityTravelInfoSection
  | SanityFeaturedToursSection
  | SanityWhatsIncludedSection

export type SanityTourPageSectionNames = SanityTourPageSection["_type"]

export type SanityTourPage = {
  _type: "tour_page"
  _id: string
  _key: string
  destination?: {
    _type: "reference"
    _id: string
    _ref: string
  } & SanityDestinationPage
  meta_data?: SanityMetaData
  slug?: SanitySlug
  promo_banner?: SanityPromoBanner
  hero_section?: {
    title?: SanityLocaleString
    image?: SanityImage
  }
  timeline?: {
    timeline?: SanityTourTimeline
    disabled?: SanityTimeline[]
  }
  price_override?: {
    timeline?: SanityTimeline
    price?: SanityPrice
  }[]
  payment?: {
    room_options?: {
      title?: SanityLocaleString
      description?: SanityLocaleString
      image?: SanityImage
      rating?: number
      price?: SanityPrice
    }[]
    room_sharing_options?: {
      title?: SanityLocaleString
      description?: SanityLocaleString
      image?: SanityImage
      price?: SanityPrice
    }[]
    extras?: {
      _key: string
      city_name?: SanityLocaleString
      count?: number
      visits?: {
        _key: string
        title?: SanityLocaleString
        description?: SanityLocaleString
        image?: SanityImage
        price?: SanityPrice
      }[]
    }[]
  }
  overview_card?: {
    duration?: SanityLocaleString
    countries?: number
    cities?: number
    rating?: number
    about?: SanityLocaleString
    price?: SanityPrice
    cta_button?: SanityLinkButton
    cta_helper_text?: SanityLocaleString
  }
  sections?: SanityTourPageSection[]
}